requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
pydantic
